added CompoundCoordinateTransformations.py for coordinate transformations that require AstrometryUtils.py
moved raDecFromAltAz and altAzPaFromRaDec to that file
