+++ python/lsst/sims/utils/CoordinateTransformations.py
@@ -7,8 +7,6 @@ __all__ = ["_galacticFromEquatorial", "galacticFromEquatorial",
            "sphericalFromCartesian", "cartesianFromSpherical",
            "rotationMatrixFromVectors",
            "equationOfEquinoxes", "calcGmstGast", "calcLmstLast",
-           "_altAzPaFromRaDec", "altAzPaFromRaDec",
-           "_raDecFromAltAz", "raDecFromAltAz",
            "_getRotSkyPos", "getRotSkyPos",
            "_getRotTelPos", "getRotTelPos",
            "haversine",
@@ -278,233 +276,6 @@ def calcGmstGast(mjd):
     gast %= 24.0
 
     return gmst, gast
-
-
-def altAzPaFromRaDec(ra, dec, lon, lat, mjd):
-    """
-    Convert RA, Dec, longitude, latitude and MJD into altitude, azimuth
-    and parallactic angle using PALPY
-
-    @param [in] ra is RA in degrees.  Can be a numpy array or a single value.
-
-    @param [in] dec is Dec in degrees.  Can be a numpy array or a single value.
-
-    @param [in] lon is the longitude of the observer in degrees
-    (positive east of the prime meridian).  Must be a single value.
-
-    @param [in] lat is the latitude of the observer in degrees
-    (positive north of the equator).  Must be a single value.
-
-    @param [in] mjd is the Universal Time expressed as an MJD.
-    Can be a numpy array or a single value.  If a numpy array, should have the
-    same number of entries as ra and dec.  In this case,
-    each mjd will be applied to its corresponding ra, dec pair.
-
-    @param [out] altitude in degrees
-
-    @param [out] azimuth in degrees
-
-    @param [out] parallactic angle in degrees
-    """
-
-    alt, az, pa = _altAzPaFromRaDec(numpy.radians(ra), numpy.radians(dec),
-                                    numpy.radians(lon), numpy.radians(lat),
-                                    mjd)
-
-    return numpy.degrees(alt), numpy.degrees(az), numpy.degrees(pa)
-
-
-
-def _altAzPaFromRaDec(raRad, decRad, longRad, latRad, mjd):
-    """
-    Convert RA, Dec, longitude, latitude and MJD into altitude, azimuth
-    and parallactic angle using PALPY
-
-    @param [in] raRad is RA in radians.  Can be a numpy array or a single value.
-
-    @param [in] decRad is Dec in radians.  Can be a numpy array or a single value.
-
-    @param [in] longRad is the longitude of the observer in radians
-    (positive east of the prime meridian).  Must be a single value.
-
-    @param [in] latRad is the latitude of the observer in radians
-    (positive north of the equator).  Must be a single value.
-
-    @param [in] mjd is the Universal Time expressed as an MJD.
-    Can be a numpy array or a single value.  If a numpy array, should have the
-    same number of entries as ra and dec.  In this case,
-    each mjd will be applied to its corresponding ra, dec pair.
-
-    @param [out] altitude in radians
-
-    @param [out] azimuth in radians
-
-    @param [out] parallactic angle in radians
-    """
-
-    if isinstance(longRad, numpy.ndarray):
-        raise RuntimeError('cannot pass numpy array of longitudes to altAzPaFromRaDec')
-
-    if isinstance(latRad, numpy.ndarray):
-        raise RuntimeError('cannot pass numpy array of latitudes to altAzPaFromRaDec')
-
-    raIsArray = False
-    decIsArray = False
-    mjdIsArray = False
-    if isinstance(raRad, numpy.ndarray):
-        raIsArray = True
-
-    if isinstance(decRad, numpy.ndarray):
-        decIsArray = True
-
-    if isinstance(mjd, numpy.ndarray):
-        mjdIsArray = True
-
-    if raIsArray and not decIsArray:
-        raise RuntimeError('passed numpy array of RA to altAzPaFromRaDec; but only one Dec')
-
-    if decIsArray and not raIsArray:
-        raise RuntimeError('passed numpy array of Dec to altAzPaFromRaDec; but only one RA')
-
-    if raIsArray and decIsArray and len(raRad) != len(decRad):
-        raise RuntimeError('in altAzPaFromRaDec length of RA numpy array does not match length of Dec numpy array')
-
-    if mjdIsArray and not raIsArray:
-        raise RuntimeError('passed numpy array of mjd to altAzPaFromRaDec; but only one RA, Dec')
-
-    if mjdIsArray and len(mjd) != len(raRad):
-        raise RuntimeError('in altAzPaFromRaDec length of mjd numpy array is not the same as length of RA numpy array')
-
-    lst = calcLmstLast(mjd, longRad)
-    last = lst[1]
-    haRad = numpy.radians(last*15.0) - raRad
-
-    if isinstance(haRad, numpy.ndarray):
-        az, azd, azdd, \
-        alt, altd, altdd, \
-        pa, pad, padd = palpy.altazVector(haRad, decRad, latRad)
-    else:
-        az, azd, azdd, \
-        alt, altd, altdd, \
-        pa, pad, padd = palpy.altaz(haRad, decRad, latRad)
-
-    return alt, az, pa
-
-
-def raDecFromAltAz(alt, az, lon, lat, mjd):
-    """
-    Convert altitude and azimuth to RA and Dec
-
-    @param [in] alt is the altitude in degrees.  Can be a numpy array or a single value.
-
-    @param [in] az is the azimuth in degrees.  Cant be a numpy array or a single value.
-
-    @param [in] lon is the observatory longitude in degrees
-    (positive east of the prime meridian).  Must be a single value.
-
-    @param [in] lat is the latitude in degrees
-    (positive north of the equator).  Must be a single value.
-
-    @param [in] mjd is the Universal Time expressed as an MD.
-    Can be a numpy array or a single value.  If a single value, must have the same length as
-    the numpy array of alt and az.  In this case, each MJD will be associated with the corresponding
-    alt, az pair.
-
-    @param [out] RA in degrees
-
-    @param [out] Dec in degrees
-
-    Note: This method is only accurate to within 0.01 arcsec near azimuth = 0 or pi
-    """
-
-    ra, dec = _raDecFromAltAz(numpy.radians(alt), numpy.radians(az),
-                              numpy.radians(lon), numpy.radians(lat), mjd)
-
-    return numpy.degrees(ra), numpy.degrees(dec)
-
-
-def _raDecFromAltAz(altRad, azRad, longRad, latRad, mjd):
-    """
-    Convert altitude and azimuth to RA and Dec
-
-    @param [in] altRad is the altitude in radians.  Can be a numpy array or a single value.
-
-    @param [in] azRad is the azimuth in radians.  Cant be a numpy array or a single value.
-
-    @param [in] longRad is the observatory longitude in radians
-    (positive east of the prime meridian).  Must be a single value.
-
-    @param [in] latRad is the latitude in radians
-    (positive north of the equator).  Must be a single value.
-
-    @param [in] mjd is the Universal Time expressed as an MD.
-    Can be a numpy array or a single value.  If a single value, must have the same length as
-    the numpy array of alt and az.  In this case, each MJD will be associated with the corresponding
-    alt, az pair.
-
-    @param [out] RA in radians
-
-    @param [out] Dec in radians
-
-    Note: This method is only accurate to within 0.01 arcsec near azimuth = 0 or pi
-    """
-    if isinstance(longRad, numpy.ndarray):
-        raise RuntimeError('cannot pass a numpy array of longitudes to raDecFromAltAz')
-
-    if isinstance(latRad, numpy.ndarray):
-        raise RuntimeError('cannot pass a numpy array of latitudes to raDecFromAltAz')
-
-    mjdIsArray = False
-    altIsArray = False
-    azIsArray = False
-    if isinstance(mjd, numpy.ndarray):
-        mjdIsArray = True
-
-    if isinstance(altRad, numpy.ndarray):
-        altIsArray = True
-
-    if isinstance(azRad, numpy.ndarray):
-        azIsArray = True
-
-    if altIsArray and not azIsArray:
-        raise RuntimeError('passed a numpy array of alt to raDecFromAltAz, but only one az')
-
-    if azIsArray and not altIsArray:
-        raise RuntimeError('passed a numpy array of az to raDecFromAltAz, but only one alt')
-
-    if azIsArray and altIsArray and len(altRad)!=len(azRad):
-        raise RuntimeError('in raDecFromAltAz, length of alt numpy array does not match length of az numpy array')
-
-    if mjdIsArray and not azIsArray:
-        raise RuntimeError('passed a numpy array of mjd to raDecFromAltAz, but only one alt, az pair')
-
-    if mjdIsArray and len(mjd) != len(azRad):
-        raise RuntimeError('in raDecFromAltAz length of mjd numpy array does not match length of az numpy array')
-
-    lst = calcLmstLast(mjd, longRad)
-    last = lst[1]
-    sinAlt = numpy.sin(altRad)
-    cosLat = numpy.cos(latRad)
-    sinLat = numpy.sin(latRad)
-    decRad = numpy.arcsin(sinLat*sinAlt+ cosLat*numpy.cos(altRad)*numpy.cos(azRad))
-    costheta = (sinAlt - numpy.sin(decRad)*sinLat)/(numpy.cos(decRad)*cosLat)
-    if altIsArray:
-        haRad0 =  numpy.arccos(costheta)
-        # Make sure there were no NaNs
-        nanSpots = numpy.where(numpy.isnan(haRad0))[0]
-        if numpy.size(nanSpots) > 0:
-            haRad0[nanSpots] = 0.5*numpy.pi*(1.0-numpy.sign(costheta[nanSpots]))
-    else:
-        haRad0 = numpy.arccos(costheta)
-        if numpy.isnan(haRad0):
-            if numpy.sign(costheta)>0.0:
-                haRad0 = 0.0
-            else:
-                haRad0 = numpy.pi
-
-    haRad = numpy.where(numpy.sin(azRad)>=0.0, -1.0*haRad0, haRad0)
-    raRad = numpy.radians(last*15.) - haRad
-    return raRad, decRad
 
 
 def getRotSkyPos(ra, dec, lon, lat, mjd, rotTel):
